api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
requests
chromadb
pdfplumber
pytesseract
Pillow
python-multipart
